* [jsfm] renamed runtime/frameworks into runtime/config
